+++ ui/src/collections/collectionUpsertModal.js
@@ -7,6 +7,7 @@ import { collectionViewQueryTab } from "./collectionViewQueryTab";
 window.app = window.app || {};
 window.app.modals = window.app.modals || {};
 
+// @todo consider adding an option for setting the initial open tab
 window.app.modals.openCollectionUpsert = function(collection = {}, modalSettings = {
     // base modal events
     onbeforeopen: null, // function(el) {},