+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Limit with ellipsis long `url` field values.
 
+- Readded the "fullscreen" `editor` field option and preloaded the TinyMCE component for slightly faster initial rendering ([#7746](https://github.com/pocketbase/pocketbase/issues/7746)).
+
 
 ## v0.39.4